Task Scheduler v0.3
New update!

Removed lots of code, and redesigned all tasks and functions to work more efficiently.

I also added the option for all tasks to return either a "SUCCESS"- or "FAILED" message. This can be used to determine whether the task you just created, changed, or deleted, indeed were that.